+++ ohsome_quality_api/ohsomedb/__init__.py
@@ -1,4 +1,4 @@
-from .contributions import (
+from .requests import (
     attribute_completeness,
     contributions,
     density,
